Single-precision is part of core
It has been objected that while the other floating-point precision
capabilities always have something next to them (either n/a or the
extension through which they are supported), single-precision doesn't.
Since it's always supported because it's part of core, just mark it as
such, to make it consistent with the other, and with things such as the
support for device partition.