Memory suffix tuning
A cl_ulong can represent up to 16EiB, so let's extend the suffix table
up to that, and in fact beyond, down the YiB.

It is now technically impossible for strbuf_mem to move beyond the end
of the array (cf. Debian Bug #848177) since no memory amount that can be
represented will trigger the off-by-one, but let's fix the logic anyway.