Different hack for function vs data pointer aliasing
Apparently older gcc versions (pre-4.6 at the very least) have a
different (stricter?) aliasing detection logic, which prevents the hack
we use to cast the return value of clGetExtensionFunctionAddress into a
proper function pointer.

Simply disabling the warning in the relevant area doesn't seem to be
sufficient for some reason, but this can be worked around by going the
indirect route (i.e. casting a pointer-to-pointer rather than the
original pointers we're interested in) in a different way.

Fixes GH issue #33.