Tentative fix for multiarch on Windows
The MS linker apparently stops at the first instance it finds of each
file, so we can't specify the path for both the 32- and 64-bit
libraries on the same line.

Split the library path specification by architecture, and detect which
one should be used.

(The detection is most definitely flaky for the time being, but it
works on my build VM at least.)