Tentative support for cl_intel_simulataneous_sharing
By a more careful reading of the specification and some tests on
machines that seem to support it, it emerges that the values are
(related to) the ones available as extra context properties in the
various cl_*_sharing extensions. Fill up a table of names accordingly.
This may be sufficient.