Improve Android support
Extend the default library path to add the vendor libraries
(which is where `libOpenCL.so` is typically found),
and include some documentation on how to build clinfo on Android in a
Termux session.

Running still requires a library search path override due to the
way the Bionic runtime linker works (no DT_RPATH or DT_RUNPATH support).

Android is not autodetected (`uname -s` returns `Linux`) so it has to be
manually specified by the user. (A possible idea would be to look at the
output of `uname -o` in Linux, since this does mention Android.)